chore: move test folder to root
