--- src/setupJest.js
@@ -1,3 +1,0 @@
-import { Testing } from 'cdktf';
-
-Testing.setupJest();
