+++ test/cdktf.fixture.json
@@ -1,0 +1,8 @@
+{
+    "app": "npx ts-node main.ts",
+    "language": "typescript",
+    "projectId": "12345678-1234-1234-1234-123456789012",
+    "sendCrashReports": "false",
+    "terraformProviders": [],
+    "terraformModules": []
+}
