+++ test/setupJest.js
@@ -1,0 +1,3 @@
+import { Testing } from 'cdktf';
+
+Testing.setupJest();
